Added terminal clause meta scope
The symbol identification was picking up the closing name on several constructs when there wasn't a secondary clause.  For example entity would show up twice.  To combat this I put in a scope on the terminal line to help identify and remove it from the list.  In other instaces (processes) it is sufficient to identify the declaration from the body.