Issue #57: Added passive concurrent statement lexical support to `entity`
The entity scoping would not handle the passive block.  I've added support for this, along with a few other statements like assertion and procedure subprogram calls (very generic -- no other way to detect this).  Lost the ability to detect non-matching identifiers after the `begin` as is usual when using the syntax `set` command.