Issue #61: Fixed multiple identifiers per line bug
The original syntax definition worked fine for single identifiers but as soon as there were multiple identifiers in a signal declaration, it would not match.  Trying to extend the same style to multiple capture groups did not work due to a limitation in Sublime's lexing state machine.  Modified the style to look for the initial declarative word, then matching identifiers singly after that (along with punctuation) up until the colon and then looked for the semicolon.  This also ended up impacting constant and variable declarations as well.