Issue #63: Assignment with a function in the index.
Signal assignment with a function (like `to_integer`) in the index area would not identify as a signal assignment.  Using a lookahead to try to identify this.  Might be able to replace other assignments like this but for now just left as a final resort on signal assignment until I have time to do further testing and experimentation.