Issues #74 #75: Fixing syntax definition
Two changes here.  first there was what seemed to be a small issue with a procedure arguments for a procedure not being scoped correctly.  This turned into a need to separate out all library subprograms into functions and procedures and correctly identify the end of both.  Secondarily, the interior of a generate statement is very complex and thus added some includes in there to help scope that correctly.