Issue #81 More hooks no solution
Made an initial stab at trying to see if I could get follow on comment lines to an inline comment to align properly but this breaks many things (not the least of which is breaking "box" comments).  Leaving the hooks in there though for the moment just in case I have a brainstorm on how to make this work.
Also added some comments because align_block_on_re was not well commented and I needed to remind myself how it worked.