Masking Fix
I tend to use "----" a lot in testbenches and discovered that my comment masking would incorrectly mask this which would lead to problems with indentation because it'd break the string.  So always masking strings then comments, and altered restore to work matches in reverse for the occasional case of a string literal inside of a comment.