Scoping Experiment
I wondered if getting the scope at the first character rather than the first line would produce better results.  It does solve a few issues like `when <choice> =>` aligning `=>` with something it shouldn't, but then it has some funny results also with aligning `when` in long mux assignments, so I opted to keep it the way it was, but left in the hooks for later if I find a better way to use it.